LayoutTests/imported/w3c/web-platform-tests/css/css-ui/animation/outline-width-composition-expected.txt


FAIL Compositing: property <outline-width> underlying [50px] from add [100px] to add [200px] at (-0.3) should be [120px] assert_equals: expected "120px " but got "70px "
FAIL Compositing: property <outline-width> underlying [50px] from add [100px] to add [200px] at (0) should be [150px] assert_equals: expected "150px " but got "100px "
FAIL Compositing: property <outline-width> underlying [50px] from add [100px] to add [200px] at (0.5) should be [200px] assert_equals: expected "200px " but got "150px "
FAIL Compositing: property <outline-width> underlying [50px] from add [100px] to add [200px] at (1) should be [250px] assert_equals: expected "250px " but got "200px "
FAIL Compositing: property <outline-width> underlying [50px] from add [100px] to add [200px] at (1.5) should be [300px] assert_equals: expected "300px " but got "250px "
FAIL Compositing: property <outline-width> underlying [100px] from add [10px] to add [2px] at (-0.5) should be [114px] assert_equals: expected "114px " but got "14px "
FAIL Compositing: property <outline-width> underlying [100px] from add [10px] to add [2px] at (0) should be [110px] assert_equals: expected "110px " but got "10px "
FAIL Compositing: property <outline-width> underlying [100px] from add [10px] to add [2px] at (0.5) should be [106px] assert_equals: expected "106px " but got "6px "
FAIL Compositing: property <outline-width> underlying [100px] from add [10px] to add [2px] at (1) should be [102px] assert_equals: expected "102px " but got "2px "
FAIL Compositing: property <outline-width> underlying [100px] from add [10px] to add [2px] at (1.5) should be [98px] assert_equals: expected "98px " but got "- 2px "
FAIL Compositing: property <outline-width> underlying [10em] from add [100px] to add [20em] at (-0.3) should be [calc(130px + 4em)] assert_equals: expected "194px " but got "34px "
FAIL Compositing: property <outline-width> underlying [10em] from add [100px] to add [20em] at (0) should be [calc(100px + 10em)] assert_equals: expected "260px " but got "100px "
FAIL Compositing: property <outline-width> underlying [10em] from add [100px] to add [20em] at (0.5) should be [calc(50px + 20em)] assert_equals: expected "370px " but got "210px "
FAIL Compositing: property <outline-width> underlying [10em] from add [100px] to add [20em] at (1) should be [30em] assert_equals: expected "480px " but got "320px "
FAIL Compositing: property <outline-width> underlying [10em] from add [100px] to add [20em] at (1.5) should be [calc(-50px + 40em)] assert_equals: expected "590px " but got "430px "
FAIL Compositing: property <outline-width> underlying [50px] from add [100px] to replace [200px] at (-0.3) should be [135px] assert_equals: expected "135px " but got "70px "
FAIL Compositing: property <outline-width> underlying [50px] from add [100px] to replace [200px] at (0) should be [150px] assert_equals: expected "150px " but got "100px "
FAIL Compositing: property <outline-width> underlying [50px] from add [100px] to replace [200px] at (0.5) should be [175px] assert_equals: expected "175px " but got "150px "
PASS Compositing: property <outline-width> underlying [50px] from add [100px] to replace [200px] at (1) should be [200px]
FAIL Compositing: property <outline-width> underlying [50px] from add [100px] to replace [200px] at (1.5) should be [225px] assert_equals: expected "225px " but got "250px "
FAIL Compositing: property <outline-width> underlying [thick] from add [11px] to add [thin] at (-0.3) should be [19px] assert_equals: expected "19px " but got "14px "
FAIL Compositing: property <outline-width> underlying [thick] from add [11px] to add [thin] at (0) should be [16px] assert_equals: expected "16px " but got "11px "
FAIL Compositing: property <outline-width> underlying [thick] from add [11px] to add [thin] at (0.5) should be [11px] assert_equals: expected "11px " but got "6px "
FAIL Compositing: property <outline-width> underlying [thick] from add [11px] to add [thin] at (1) should be [6px] assert_equals: expected "6px " but got "1px "
FAIL Compositing: property <outline-width> underlying [thick] from add [11px] to add [thin] at (1.5) should be [1px] assert_equals: expected "1px " but got "- 4px "
FAIL Compositing: property <outline-width> underlying [thick] from add [11px] to add [thin] at (2) should be [0px] assert_equals: expected "0px " but got "- 9px "
